# openapi.yaml
openapi: 3.0.3
info:
  title: računAI API
  version: 1.0.0
  description: Tenant JWT REST ugovor za auth, documents, banking, partners, finance,
    assets, purchasing i tax. Fiscal/intermediary i AS4 nisu dio ove sheme. openapi.yaml
    je generirani artefakt — ne uređivati ručno.
paths:
...
  /api/purchasing/eracun/inbound-import/:
    post:
      operationId: purchasing_eracun_inbound_import
      description: Import gateway inbox documents into draft expenses. No provider
        round-trip.
      tags:
      - purchasing
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EracunInboundImportResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
  /api/purchasing/eracun/inbox-refresh/:
    post:
      operationId: purchasing_eracun_inbox_refresh
      description: Ask the gateway to reconcile against the provider. Imports nothing.
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        required: true
      tags:
      - purchasing
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EracunInboxRefreshResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
  /api/purchasing/expense-categories/:
    get:
      operationId: purchasing_expense_categories_list
      tags:
      - purchasing
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseCategoryList'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/purchasing/expense-categories/{id}/:
    patch:
      operationId: purchasing_expense_categories_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - purchasing
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedExpenseCategoryPatchRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedExpenseCategoryPatchRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedExpenseCategoryPatchRequest'
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseCategory'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/purchasing/expenses/{id}/eracun-rejection/:
    post:
      operationId: purchasing_expenses_eracun_rejection
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        required: true
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - purchasing
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EracunRejectionRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EracunRejectionRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EracunRejectionRequestRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EracunRejectionResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
  /api/purchasing/invoices/import/:
    post:
      operationId: purchasing_invoice_imports_create
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        required: true
      tags:
      - purchasing
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      security:
      - bearerAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingInvoiceImport'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Neprocesibilan zahtjev (validacija write operacije)
  /api/purchasing/invoices/import/{id}/:
    get:
      operationId: purchasing_invoice_imports_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - purchasing
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingInvoiceImport'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
  /api/purchasing/invoices/import/{id}/apply-partner-updates/:
    post:
      operationId: purchasing_invoice_imports_apply_partner_updates
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - purchasing
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingInvoiceImport'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
  /api/purchasing/invoices/import/{id}/confirm/:
    post:
      operationId: purchasing_invoice_imports_confirm
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - purchasing
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmInvoiceImportRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ConfirmInvoiceImportRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ConfirmInvoiceImportRequest'
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingInvoiceImport'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Nevaljani upit / ValidationError
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Neprocesibilan zahtjev (validacija write operacije)
...
  /api/purchasing/invoices/import/{id}/discard/:
    post:
      operationId: purchasing_invoice_imports_discard
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - purchasing
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingInvoiceImport'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                Unauthorized:
                  value:
                    detail: Authentication credentials were not provided.
          description: Nedostaje ili je nevaljan Bearer token
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: 'Nije pronađeno: missing resource, cross-tenant ID, ili autenticiran
            korisnik bez prava (namjerno 404, ne 403)'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasingConflict'
          description: ''
...
components:
  schemas:
...
    EracunInboundImportResponse:
      type: object
      properties:
        scanned:
          type: integer
        imported:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        remaining_importable:
          type: integer
        errors:
          type: array
          items:
            $ref: '#/components/schemas/EracunImportError'
      required:
      - errors
      - failed
      - imported
      - remaining_importable
      - scanned
      - skipped
    EracunInboxRefreshResponse:
      type: object
      properties:
        reconciliation_id:
          type: string
        status:
          type: string
        detail:
          type: string
      required:
      - detail
      - reconciliation_id
      - status
...
    EracunRejectionRequestRequest:
      type: object
      properties:
        reason_code:
          $ref: '#/components/schemas/ReasonCodeEnum'
        reason_text:
          type: string
          maxLength: 2000
      required:
      - reason_code
...
    ErrorDetail:
      type: object
      description: 'DRF-style error body: ``{"detail": ...}`` where detail may be
        str or object.'
      properties:
        detail: {}
      required:
      - detail
...
    ExpenseCategoryList:
      type: object
      properties:
        count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseCategory'
      required:
      - count
      - results
...
    PatchedExpenseCategoryPatchRequest:
      type: object
      properties:
        default_account_id:
          type: integer
          nullable: true
        default_cost_center_id:
          type: integer
          nullable: true
...
    PurchasingConflict:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
      required:
      - code
      - detail
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...